% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping.R
\docType{methods}
\name{add_auto_group}
\alias{add_auto_group}
\alias{add_auto_group,Rscexv-method}
\title{description of function group_1D}
\usage{
add_auto_group(dataObj, groupData, groupName)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{groupData}{a vector with group ids in the same order as the samples table}

\item{groupName}{the new name of the group}
}
\description{
Checks the grouping data and does 
NOT add the grouping if it is not changing the order of the samples
}

