% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCExV_Functions_analyse.data.R
\docType{methods}
\name{analyse.data}
\alias{analyse.data}
\alias{analyse.data,Rscexv-method}
\title{description of function analyse.data}
\usage{
analyse.data(obj, onwhat = "Expression", groups.n, cmethod,
  clusterby = "MDS", ctype = "hierarchical clust", beanplots = TRUE,
  move.neg = FALSE, plot.neg = TRUE, useGrouping = NULL, plotsvg = 0,
  mds.type = "PCA", ...)
}
\arguments{
\item{obj}{the Rscexv object}

\item{onwhat}{cluster on weither Expression of FACS data default='Expression'}

\item{groups.n}{the number of groups to create}

\item{cmethod}{clustering method}

\item{clusterby}{which dataset to cluter on (raw or MDS) default='MDS'}

\item{ctype}{the clzutering type default = 'hierarchical clust'}

\item{beanplots}{create beanplots or violinplots}

\item{move.neg}{should the not expressed value be as set to the minimum expression value?}

\item{plot.neg}{show the not expressing samples in the bean/violinplots?}

\item{useGrouping}{do not calculate a new grouping - use this column in the samples table (default=NULL)}

\item{plotsvg}{create svg figures in addition to the png ones (default=0)}

\item{mds.type}{which mds algorithm to use (PCA, LLE, ISOMAP or ZIFA) default = PCA}
}
\description{
This function implements the backend functionallity of the SCExV analyse page.
}

