% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_01_class.R
\docType{data}
\name{analyzed}
\alias{analyzed}
\alias{data}
\title{This is the analyzed data used for downstream testing}
\format{
Rscexv object
}
\usage{
data(analyzed)
}
\description{
This file can be re-created running the test scripts.
}
\keyword{datasets}
