% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvalues_boot_p_value.R
\docType{methods}
\name{boot_p_value}
\alias{boot_p_value}
\alias{boot_p_value,Rscexv-method}
\title{description of function boot_p_value}
\usage{
boot_p_value(cmp, real_val, i)
}
\arguments{
\item{cmp}{the boot strap list of correlation values}

\item{real_val}{the correlation value}

\item{i}{the numer of bootstrap runs}
}
\description{
calculates the p value from the boot strap distribution for the lin_lang analysis
}
