% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvalues_calc.lin.lang.4_list.R
\docType{methods}
\name{calc.lin.lang.4_list}
\alias{calc.lin.lang.4_list}
\alias{calc.lin.lang.4_list,Rscexv-method}
\title{description of function calc.lin.lang.4_list}
\usage{
calc.lin.lang.4_list(l = list())
}
\arguments{
\item{l}{the list of values}
}
\description{
This function calculates the correlation between the median expression value of a group and the fraction of expression cells in the group.
}
