% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_check.dataObj.R
\docType{methods}
\name{check.dataObj}
\alias{check.dataObj}
\alias{check.dataObj,Rscexv-method}
\title{description of function check.dataObj}
\usage{
check.dataObj(PCR, FACS)
}
\arguments{
\item{PCR}{the object from read.PCR.set}

\item{FACS}{the object from read.FACS.set}
}
\description{
checks the two data frames and creates a sample and annotation table
}
