% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_checkGrouping.R
\docType{methods}
\name{checkGrouping}
\alias{checkGrouping}
\alias{checkGrouping,Rscexv-method}
\title{description of function checkGrouping}
\usage{
checkGrouping(userGroups, data = NULL)
}
\arguments{
\item{userGroups}{the grouping vector}

\item{data}{the optional Rscexv object (at the moment unused)}
}
\description{
This function checks that there is no 0 group - probably more later on
}
