% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coexpression_coexpressionMatrix.R
\docType{methods}
\name{coexpressionMatrix}
\alias{coexpressionMatrix}
\alias{coexpressionMatrix,Rscexv-method}
\title{description of function coexpressionMatrix}
\usage{
coexpressionMatrix(dataObj)
}
\arguments{
\item{dataObj}{the Rscexv object}
}
\description{
this function calculates the coexpression matrix for the expression data.
}

