% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\docType{methods}
\name{col4bean}
\alias{col4bean}
\alias{col4bean,Rscexv-method}
\title{description of function col4bean}
\usage{
col4bean(x, tic = "black")
}
\arguments{
\item{x}{TEXT MISSING}

\item{tic}{TEXT MISSING default='black'}
}
\description{
get the color information for the beans
}

