% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvalues_create_p_values.R
\docType{methods}
\name{create_p_values}
\alias{create_p_values}
\alias{create_p_values,Rscexv-method}
\title{description of function create_p_values}
\usage{
create_p_values(obj, boot = 1000, lin_lang_file = "lin_lang_stats.xls",
  sca_ofile = "Significant_genes.csv")
}
\arguments{
\item{obj}{the Rscexv object}

\item{boot}{for the boot strap approach - how many runs default= 1000}

\item{lin_lang_file}{the own stat outfile default='lin_lang_stats.xls'}

\item{sca_ofile}{the SingleCellsAssay p values outfile default="Significant_genes.csv"}
}
\description{
This is the main function called by the SCExV server.
}

