% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCExV_Functions_createDataObj.R
\docType{methods}
\name{createDataObj}
\alias{createDataObj}
\alias{createDataObj,Rscexv-method}
\title{description of function createDataObj}
\usage{
createDataObj(PCR = NULL, FACS = NULL, max.value = 40, ref.genes = NULL,
  max.ct = 25, max.control = 0, norm.function = "none",
  negContrGenes = NULL, use_pass_fail = T, ...)
}
\arguments{
\item{PCR}{the pcr data file names default=NULL}

\item{FACS}{the FACS data file names MISSING default=NULL}

\item{max.value}{an optional maximum value for all failed wells default=40}

\item{ref.genes}{if the data should be normalized to reference genes put them into this list}

\item{max.ct}{all sammples with a ct value of more than max.ct in the ref.genes is dropped}

\item{max.control}{samples are dropped if 1 (0), 2(1), ... etc ref.genes shows a ct value of more than max.ct}

\item{norm.function}{any of ( "none","mean control genes","max expression","median expression","quantile" )}

\item{negContrGenes}{which negative control genes to use (samples with expression there are dropped!)}

\item{use_pass_fail}{whether or not to use the pass_fail, information in the PCR data files.}
}
\description{
This function is the main function used by SCExV. In a script the steps should be executed one by one.
}

