% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRFgrouping_samples.R
\docType{methods}
\name{createRFgrouping_samples}
\alias{createRFgrouping_samples}
\alias{createRFgrouping_samples,Rscexv-method}
\title{description of function createRFgrouping_samples}
\usage{
createRFgrouping_samples(
  x,
  RFname,
  k = 10,
  single_res_col = paste("RFgrouping", RFname)
)
}
\arguments{
\item{x}{the Rscexv object}

\item{RFname}{the name of the RFclust.SGE object in the Rscexv data. This object has to be populized with data!}

\item{k}{the number of wanted groups ( default = 10)}

\item{single_res_col}{the new column in the samples table default= paste('RFgrouping', RFname)}
}
\description{
Create a sample grouping data from one RFclust.SGE object
}
