% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_difference.R
\docType{methods}
\name{difference}
\alias{difference}
\alias{difference,Rscexv-method}
\title{description of function difference}
\usage{
difference(x, obj)
}
\arguments{
\item{x}{TEXT MISSING}

\item{obj}{TEXT MISSING}
}
\description{
This function calculates the
}

