% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCExV_Functions_exportGroups.R
\docType{methods}
\name{exportGroups}
\alias{exportGroups}
\alias{exportGroups,Rscexv-method}
\title{description of function analyse.data}
\usage{
exportGroups(obj, file = "SCExV_Grps.txt")
}
\arguments{
\item{obj}{the Rscexv object}

\item{file}{theoutfile default='SCExV_Grps.txt'}
}
\description{
This function exports the group names necessary for the SCExV server
}

