% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_FACS.heatmap.R
\docType{methods}
\name{FACS.heatmap}
\alias{FACS.heatmap}
\alias{FACS.heatmap,Rscexv-method}
\title{description of function FACS.heatmap}
\usage{
FACS.heatmap(dataObj, ofile, title = "Heatmap", reorder = F, nmax = 500,
  hc.row = NA, ColSideColors = NA, RowSideColors = NA, width = 6,
  height = 6, margins = c(15, 10), hclustfun = function(c) {     hclust(c,
  method = "ward") }, distfun = function(x) as.dist(1 - cor(t(x), method =
  "pearson")), ...)
}
\arguments{
\item{dataObj}{TEXT MISSING}

\item{ofile}{TEXT MISSING}

\item{title}{TEXT MISSING default='Heatmap'}

\item{reorder}{if yes the data matrix will be reordered to the clusters object (default=F)}

\item{nmax}{TEXT MISSING default=500}

\item{hc.row}{TEXT MISSING default=NA}

\item{ColSideColors}{TEXT MISSING default=NA}

\item{RowSideColors}{TEXT MISSING default=NA}

\item{FACS.heatmap}{TEXT MISSING}
}
\description{
Plots the FACS heatmaps
}

