% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.on.controls.no.inv.R
\docType{methods}
\name{filter.on.controls.no.inv}
\alias{filter.on.controls.no.inv}
\alias{filter.on.controls.no.inv,Rscexv-method}
\title{description of function filter.on.controls.no.inv}
\usage{
filter.on.controls.no.inv(dataObj, ref.nms, thresh, howm)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{ref.nms}{the gene names used as positive control}

\item{thresh}{the minimum expression in each ref gene}

\item{howm}{drop if one sample has howm genes that fail the test}
}
\description{
This function filteres out samples that have an expression of less than thresh in there ref.nms(gene names)
}
