% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.unique.sample.R
\docType{methods}
\name{force.unique.sample}
\alias{force.unique.sample}
\alias{force.unique.sample,Rscexv-method}
\title{description of function force.unique.sample}
\usage{
force.unique.sample(x)
}
\arguments{
\item{x}{TEXT MISSING}
}
\description{
This function will just make sure, that all samples have uniaue sample names, even if they come from different arrays.
}
