% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_group_1D.R
\docType{methods}
\name{group_1D}
\alias{group_1D}
\alias{group_1D,Rscexv-method}
\title{description of function group_1D}
\usage{
group_1D(dataObj, gene, ranges)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{gene}{the gene the groups are based on}

\item{ranges}{the ranges for the rgoups}
}
\description{
Create grouping based on the expression of one gene
}

