% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_group_1D_worker.R
\docType{methods}
\name{group_1D_worker}
\alias{group_1D_worker}
\alias{group_1D_worker,Rscexv-method}
\title{description of function group_1D_worker}
\usage{
group_1D_worker(ma, gene, ranges)
}
\arguments{
\item{ma}{one data matrix from the Rscexv object}

\item{gene}{the gene to analyze}

\item{ranges}{the cut offs to create}
}
\description{
grouping worker for the 1D grouping.
}
