% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_group_on_strings.R
\docType{methods}
\name{group_on_strings}
\alias{group_on_strings}
\alias{group_on_strings,Rscexv-method}
\title{description of function group_on_strings}
\usage{
group_on_strings(dataObj, strings = c(), name = "GroupOnStrings")
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{strings}{the grouping strings as vector default= c()}

\item{name}{the new group name}
}
\description{
group_an_strings allows the user to group based on a string in the sample names.
}
