% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyBestGrouping.R
\docType{methods}
\name{identifyBestGrouping}
\alias{identifyBestGrouping}
\alias{identifyBestGrouping,Rscexv-method}
\title{description of function identifyBestGrouping}
\usage{
identifyBestGrouping(
  x,
  groups,
  namePrefix = "identifyBestGrouping",
  cut = 0.05
)
}
\arguments{
\item{x}{the Rscexv}

\item{groups}{the colnames (samples) that contain the grouping information (default)}

\item{namePrefix}{a common name prefix for this analysis}

\item{cut}{define the p value cut off for the test ( if too view signfican genes are detected at 0.05)}
}
\value{
A list containing the number of significant genes in each possible combination of groups and the Rscexv object containg all annotations and stats.
}
\description{
This function compares several groupings to each other.

In addition the groupings are combined in a way, that if a sample end up in group 1 in grouing A but in group 2 in groupung B

it might be something else than a sample ending up in group 1 for both.

The quality of the groupings will be accessed in two stated (1) there should not be too many small groups and

(2) the genes should be differentially expressed in these groups. The differential expression is accessed

using a straight forward anova approach (excluding the not called genes for a Rscexv object).
}
