% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_implyCloseOrder.R
\docType{methods}
\name{implyCloseOrder}
\alias{implyCloseOrder}
\alias{implyCloseOrder,Rscexv-method}
\title{description of function implyCloseOrder}
\usage{
implyCloseOrder(dataObj, groupName = NULL, hclust.method = "complete")
}
\arguments{
\item{dataObj}{the Rscexv data object}

\item{groupName}{the name of the grouping that has to be reordered}
}
\description{
This function reorderes a grouping using hclust. The grouing will simply be converted into a factor.
}

