% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{data}
\name{InbuiltDataNOFACS}
\alias{InbuiltDataNOFACS}
\title{This is the test dataset also available at the SCExV server page}
\format{Rscexv object}
\usage{
data(InbuiltDataNOFACS)
}
\description{
Please read PMID:26437766 for more details. No analysis performed.
}
\keyword{datasets}

