% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscexv_transform.R
\docType{methods}
\name{kick.expressed.negContr.samples}
\alias{kick.expressed.negContr.samples}
\alias{kick.expressed.negContr.samples,Rscexv-method}
\title{description of function kick.expressed.negContr.samples}
\usage{
kick.expressed.negContr.samples(dataObj, negContrGenes = NULL)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{negContrGenes}{which genes should be checked default=NULL}
}
\description{
IF you are sure, that a list of genes must not be expressed in cells of interest you can fore to exclude these ceells using this function.

But this is a quite bad absolutely not recommended workflow!

This function drops samples with expression different from 999 - therefore you must call this function before removing the 999 values.
}

