% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.sample.names.R
\docType{methods}
\name{match.sample.names}
\alias{match.sample.names}
\alias{match.sample.names,Rscexv-method}
\title{description of function match.sample.names}
\usage{
match.sample.names(PCR.s, FACS.s)
}
\arguments{
\item{PCR.s}{the PCR sample names}

\item{FACS.s}{the FACS sample names}
}
\description{
This function makes sure, that both the FACS and the PCR data samples are matched in the object
}
