% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.and.clus.R
\docType{methods}
\name{mds.and.clus}
\alias{mds.and.clus}
\alias{mds.and.clus,Rscexv-method}
\title{description of function mds.and.clus}
\usage{
mds.and.clus(
  dataObj,
  ...,
  clusterby = "raw",
  mds.type = "PCA",
  groups.n,
  LLEK = 2,
  cmethod = "ward.D",
  ctype = "hierarchical clust",
  onwhat = "Expression",
  useGrouping = NULL
)
}
\arguments{
\item{dataObj}{TEXT MISSING}

\item{clusterby}{cluster by raw or MDS default="raw"}

\item{mds.type}{which MDS function to use default="PCA"}

\item{groups.n}{the numer of groups to create}

\item{LLEK}{the LLEK option for the LLE and ISOMAP mds functions default=2}

\item{cmethod}{the hyclust method to use default='ward.D'}

\item{ctype}{the clustering method ('hierarchical clust' or kmeans) default='hierarchical clust'}

\item{onwhat}{cluster on FACS or Expression data default="Expression"}

\item{useGrouping}{do not calculate a new grouping - use this column in the samples table (default=NULL)}
}
\description{
this function calculates the MDS and creates the grouping if necessary.
}
