% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeComplexSampleGroupings.R
\docType{methods}
\name{mergeComplexSampleGroupings}
\alias{mergeComplexSampleGroupings}
\alias{mergeComplexSampleGroupings,Rscexv-method}
\title{description of function mergeSampleGroupings}
\usage{
mergeComplexSampleGroupings(x, g1, g2, newName)
}
\arguments{
\item{x}{the Rscexv object}

\item{g1}{the first grouping name (can be complex)}

\item{g2}{the second grouping name (can be complex too)}

\item{newName}{the name of the new variable to create}
}
\description{
Create a new group and create shaded rainbow color shemata stored in the x@usedObj[['colorRange']] variable
}
