% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_mergeSampleGroupings.R
\docType{methods}
\name{mergeSampleGroupings}
\alias{mergeSampleGroupings}
\alias{mergeSampleGroupings,Rscexv-method}
\title{description of function mergeSampleGroupings}
\usage{
mergeSampleGroupings(x, g1, g2, newName)
}
\arguments{
\item{x}{the Rscexv object}

\item{g1}{the first grouping name (can be complex)}

\item{g2}{the second grouping name (has to be simple yes/no!)}

\item{newName}{the name of the new variable to create}
}
\description{
Create a new group and create shaded rainbow color shemata stored in the x@usedObj[['colorRange']] variable
}

