% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_norm.PCR.R
\docType{methods}
\name{norm.PCR}
\alias{norm.PCR}
\alias{norm.PCR,Rscexv-method}
\title{description of function norm.PCR}
\usage{
norm.PCR(x, meth = c("none", "mean control genes", "max expression",
  "median expression", "quantile"), ctrl = NA, max.cyc)
}
\arguments{
\item{x}{the Rscexv object}

\item{meth}{The normalization method; one of "none","mean control genes","max expression","median expression","quantile" default="none"}

\item{ctrl}{the control genes default=NA}

\item{max.cyc}{this function does also invert the Ct values this is the value that is used as maximum default=NA}
}
\description{
This function normalizes the expression data
}

