% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvalues_p.lin.lang.R
\docType{methods}
\name{p.lin.lang}
\alias{p.lin.lang}
\alias{p.lin.lang,Rscexv-method}
\title{description of function p.lin.lang}
\usage{
p.lin.lang(x, groups.n, clus, n = 1000)
}
\arguments{
\item{x}{data matrix}

\item{groups.n}{the amount of groups}

\item{clus}{the grouping vector}

\item{n}{the amount of boot strap operations default=1000}
}
\description{
Calculates the linear hypothesis p value.
}

