% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_PCR.heatmap.R
\docType{methods}
\name{PCR.heatmap}
\alias{PCR.heatmap}
\alias{PCR.heatmap,Rscexv-method}
\title{description of function PCR.heatmap}
\usage{
PCR.heatmap(dataObj, ofile, reorder = F, title = "Heatmap", nmax = 4000,
  hc.row = NA, ColSideColors = NA, RowSideColors = F, width = 6,
  height = 6, margins = c(1, 11), lwid = c(1, 6), lhei = c(1, 5),
  hclustfun = function(c) {     hclust(c, method = "ward.D") },
  distfun = function(x) as.dist(1 - cor(t(x), method = "pearson")),
  Rowv = T, ...)
}
\arguments{
\item{dataObj}{TEXT MISSING}

\item{ofile}{TEXT MISSING}

\item{reorder}{if yes the data matrix will be reordered to the clusters object (default=F)}

\item{title}{TEXT MISSING default='Heatmap'}

\item{nmax}{TEXT MISSING default=4000}

\item{hc.row}{TEXT MISSING default=NA}

\item{ColSideColors}{TEXT MISSING default=NA}

\item{RowSideColors}{TEXT MISSING default=F}

\item{PCR.heatmap}{TEXT MISSING}
}
\description{
This function plots the heatmaps
}
