% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\docType{methods}
\name{plot.beans}
\alias{plot.beans}
\alias{plot.beans,Rscexv-method}
\title{description of function plot.beans}
\usage{
plot.beans(x, groups.n, clus, boot = 1000, plot.neg = TRUE, mv = -20,
  subpath = "")
}
\arguments{
\item{x}{the Rscexv object}

\item{groups.n}{TEXT MISSING}

\item{clus}{TEXT MISSING}

\item{boot}{TEXT MISSING default= 1000}

\item{plot.neg}{TEXT MISSING default=TRUE}

\item{mv}{TEXT MISSING default=-20}

\item{subpath}{the subpath for the plots (default = '')}
}
\description{
plot the beanplots
}

