% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_plot.histograms.R
\docType{methods}
\name{plot.histograms}
\alias{plot.histograms}
\alias{plot.histograms,Rscexv-method}
\title{description of function plot.histograms}
\usage{
plot.histograms(dataObj, cuts = vector("list", 1), subpath = "preprocess")
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{cuts}{the cuts are used for the 1D gene groups default=vector('list',1)}

\item{subpath}{the subpath to plot to ( default = preprocess)}
}
\description{
This function plots all genes as histograms to check whether there ar4e clear expression differences in different plates.
}

