% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_plot.violines.R
\docType{methods}
\name{plot.violines}
\alias{plot.violines}
\alias{plot.violines,Rscexv-method}
\title{description of function plot.violines}
\usage{
plot.violines(x, groups.n, clus, plot.neg = FALSE, mv = -20, subpath = "",
  names = NULL, col = NULL)
}
\arguments{
\item{x}{the Rscexv object}

\item{groups.n}{TEXT MISSING}

\item{clus}{TEXT MISSING}

\item{plot.neg}{TEXT MISSING default=FALSE}

\item{mv}{TEXT MISSING default=-20}

\item{subpath}{the path to write the figures to (default = '')}

\item{col}{a colour vector for colouring the violines (default=NULL  -> rainbow )}
}
\description{
This fucntion converts the table into a format, that can be fead to the vioplot function
}

