% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coexpression_plotcoma.R
\docType{methods}
\name{plotcoma}
\alias{plotcoma}
\alias{plotcoma,Rscexv-method}
\title{description of function plotcoma}
\usage{
plotcoma(dataObj, fname = "/CorrelationPlot")
}
\arguments{
\item{dataObj}{TEXT MISSING}

\item{fname}{TEXT MISSING default='CorrelationPlot'}
}
\description{
This function plots the correlation plot for the grouping.

It will create an svg file if the global variable plotsvg ==1.
}
