% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\docType{methods}
\name{plotDensity}
\alias{plotDensity}
\alias{plotDensity,Rscexv-method}
\title{description of function analyse.data}
\usage{
plotDensity(obj)
}
\arguments{
\item{obj}{the Rscexv object}
}
\description{
This function creates the density plot for the analysis page
}
