% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\docType{methods}
\name{plug.999}
\alias{plug.999}
\alias{plug.999,Rscexv-method}
\title{description of function plug.999}
\usage{
plug.999(x, plug = 45)
}
\arguments{
\item{x}{the Rscexv object}

\item{plug}{the replacement value (default=45)}
}
\description{
this function replaces the 999 (failed) values by the plug value.
}

