% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFgrouping.R
\docType{methods}
\name{predict.rf}
\alias{'predict.rf,Rscexv-method}
\alias{predict.rf}
\title{description of function predict.rf}
\usage{
predict.rf(x, rf, bestColname = "predicted group using random forest")
}
\arguments{
\item{x}{the single cells ngs object}

\item{rf}{the random forst model to use for the classification}

\item{bestColname}{the column name to store the results in}
}
\value{
a Rscexv object including the results and storing the RF object in the usedObj list (bestColname)
}
\description{
simple prediction of groups using a random forest trained during the bestGrouping process
}

