% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\docType{methods}
\name{quality_of_fit}
\alias{quality_of_fit}
\alias{quality_of_fit,Rscexv-method}
\title{description of function quality_of_fit}
\usage{
quality_of_fit(obj)
}
\arguments{
\item{obj}{TEXT MISSING}
}
\description{
Calculates a quality of fit
}

