% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFgrouping_qualityTest.R
\docType{methods}
\name{qualityTest}
\alias{qualityTest}
\alias{qualityTest,Rscexv-method}
\title{description of function qualityTest}
\usage{
qualityTest(x, groups = NULL, cut = 0.05, numbers = F)
}
\arguments{
\item{x}{The Rscexv object}

\item{groups}{the group name to clister the data (default=NULL)}

\item{cut}{the p value cut off (BenjaminHochberg corrected; default=0.05)}

\item{numbers}{if true return amount of significant genes, not the names}
}
\value{
A list containing the updated 'x' Rscexv object and res - a vector of pasted gene names that turned out to be significantly different.
}
\description{
This function calculates an anova test for the groupCol and all data.
}

