% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.FACS.R
\docType{methods}
\name{read.FACS}
\alias{read.FACS}
\alias{read.FACS,Rscexv-method}
\title{description of function read.FACS}
\usage{
read.FACS(fname)
}
\arguments{
\item{fname}{the file containg the INDEX data}
}
\description{
THis function reads one INDEX data file.
}
