% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles.R
\docType{methods}
\name{read.FACS.set}
\alias{read.FACS.set}
\alias{read.FACS.set,Rscexv-method}
\title{description of function read.FACS.set}
\usage{
read.FACS.set(fnames)
}
\arguments{
\item{fnames}{TEXT MISSING}
}
\description{
reads a whole set of FACS data files
}

