% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles_read.PCR.R
\docType{methods}
\name{read.PCR}
\alias{read.PCR}
\alias{read.PCR,Rscexv-method}
\title{description of function read.PCR}
\usage{
read.PCR(fname, use_pass_fail = T)
}
\arguments{
\item{fname}{the fliename to read}

\item{use_pass_fail}{exclude data from failed wells default=T}
}
\description{
reads one fluidigm data file
}
