% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.PCR.heatmap.R
\docType{methods}
\name{read.PCR.heatmap}
\alias{read.PCR.heatmap}
\alias{read.PCR.heatmap,Rscexv-method}
\title{description of function read.PCR.heatmap}
\usage{
read.PCR.heatmap(fname, use_pass_fail = T)
}
\arguments{
\item{fname}{TEXT MISSING}

\item{use_pass_fail}{TEXT MISSING default=T}
}
\description{
reads one fluidigm data file in heatmap format
}
