% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles_read.PCR.set.R
\docType{methods}
\name{read.PCR.set}
\alias{read.PCR.set}
\alias{read.PCR.set,Rscexv-method}
\title{description of function read.PCR.set}
\usage{
read.PCR.set(fnames, use_pass_fail = T)
}
\arguments{
\item{fnames}{a list of file names}

\item{use_pass_fail}{whether to remove values marked as failed in the files (default=T)}
}
\description{
reads a whole set of PCR data files
}

