% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_regroup.R
\docType{methods}
\name{regroup}
\alias{regroup}
\alias{regroup,Rscexv-method}
\title{description of function regroup}
\usage{
regroup(dataObj, group2sample = list(`1` = c("Sample1", "Sample2")),
  name = NULL)
}
\arguments{
\item{dataObj}{TEXT MISSING}

\item{group2sample}{TEXT MISSING default= list ( '1' = c( 'Sample1'}

\item{groupName}{which group to re-group}
}
\description{
This function is used to re-group the data from within SCExV - there should be easier ways in R.
}

