% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscexv_transform_remove.FACS.genes.R
\docType{methods}
\name{remove.FACS.genes}
\alias{remove.FACS.genes}
\alias{remove.FACS.genes,Rscexv-method}
\title{description of function remove.FACS.genes}
\usage{
remove.FACS.genes(dataObj, ids)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{ids}{the gene ids to remove}
}
\description{
this function removes a list of FACS genes from the analysis
}
