% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.genes.R
\docType{methods}
\name{remove.genes}
\alias{remove.genes}
\alias{remove.genes,Rscexv-method}
\title{description of function remove.genes}
\usage{
remove.genes(dataObj, ids)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{ids}{the gene ids to remove}
}
\description{
this function removes a list of PCR genes from the analysis
}
