% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscexv_transform_remove.samples.R
\docType{methods}
\name{remove.samples}
\alias{remove.samples}
\alias{remove.samples,Rscexv-method}
\title{description of function remove.samples}
\usage{
remove.samples(dataObj, ids)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{ids}{which samples (ids!) to remove}
}
\description{
Remove samples by id.
}
