% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscexv_transform_reorder.genes.R
\docType{methods}
\name{reorder.genes}
\alias{reorder.genes}
\alias{reorder.genes,Rscexv-method}
\title{description of function remove.genes}
\usage{
reorder.genes(dataObj, column)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{column}{the annotation column to reorder on}
}
\description{
this function reorderes the Rscexv object based on a column in the annotation table (e.g. for plotting)
}

