% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscexv_transform.R
\docType{methods}
\name{reorder.samples}
\alias{reorder.samples}
\alias{reorder.samples,Rscexv-method}
\title{description of function remove.genes}
\usage{
reorder.samples(dataObj, column)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{column}{the samples column to reorder on}
}
\description{
this function reorderes the Rscexv object based on a column in the samples table (e.g. for plotting)
}

