% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Rscexv.R
\docType{methods}
\name{Rscexv}
\alias{Rscexv}
\alias{Rscexv,Rscexv-method}
\title{description of function createDataObj}
\usage{
Rscexv(PCR = NULL, FACS = NULL, use_pass_fail = T)
}
\arguments{
\item{PCR}{the pcr data file names default=NULL}

\item{FACS}{the FACS data file names MISSING default=NULL}

\item{use_pass_fail}{whether or not to use the pass_fail, information in the PCR data files.}
}
\description{
create a new Rscexv object from the data files.
}
