\name{Rscexv-package}
\alias{Rscexv-package}
\alias{Rscexv}
\docType{package}
\title{
\packageTitle{Rscexv}
}
\description{
\packageDescription{Rscexv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rscexv}
\packageIndices{Rscexv}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Rscexv}

Maintainer: \packageMaintainer{Rscexv}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
