% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{Rscexv}
\alias{Rscexv}
\title{Rscexv}
\description{
This S4 class implements the SCExV backend logics
}
\section{Slots}{

\describe{
\item{\code{data}}{a data.frame containing the expression values for each gene x sample (gene = row)}

\item{\code{facs}}{a data.frame containing all the FACS data}

\item{\code{samples}}{a data.frame describing the columnanmes in the data column}

\item{\code{annotation}}{a data.frame describing the rownames in the data rows}

\item{\code{outpath}}{the default outpath for the plots and tables from this package}

\item{\code{name}}{the name for this package (all filesnames contain that)}

\item{\code{zscored}}{genes are normalized?}

\item{\code{snorm}}{samples normalized?}

\item{\code{normFunct}}{which PCR norm function has been called}

\item{\code{rownamescol}}{the column name in the annotation table that represents the rownames of the data table}

\item{\code{sampleNamesCol}}{the column name in the samples table that represents the colnames of the data table}

\item{\code{usedObj}}{here a set of used and probably lateron important objects can be saved. Be very carful using any of them!}
}}

