% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess_scale.FACS.data.R
\docType{methods}
\name{scale.FACS.data}
\alias{scale.FACS.data}
\alias{scale.FACS.data,Rscexv-method}
\title{description of function scale.FACS.data}
\usage{
scale.FACS.data(dataObj)
}
\arguments{
\item{dataObj}{TEXT MISSING}
}
\description{
This function applies a simple log10 onto all expression data.
}
