% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.filter.R
\docType{methods}
\name{sd.filter}
\alias{sd.filter}
\alias{sd.filter,Rscexv-method}
\title{description of function sd.filter}
\usage{
sd.filter(dataObj)
}
\arguments{
\item{dataObj}{the Rscexv object}
}
\description{
This function is removing cells and genes that do not show any variation in the data.
}
