% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileLocks.R
\docType{methods}
\name{set.lock}
\alias{set.lock}
\alias{set.lock,Rscexv-method}
\title{description of function set_lock}
\usage{
set.lock(filename, msg = "no mgs")
}
\arguments{
\item{filename}{the locked file}

\item{msg}{A message that should be written into the log file ( default 'no mgs')}
}
\description{
set a lock for a file (threading)
}
