% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,Rscexv-method}
\title{description of function show}
\usage{
\S4method{show}{Rscexv}(object)
}
\arguments{
\item{object}{the Rscexv object}
}
\value{
nothing
}
\description{
print the Rscexv
}
