% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvalues_SingleCellAssay_Pvalues.R
\docType{methods}
\name{SingleCellAssay_Pvalues}
\alias{SingleCellAssay_Pvalues}
\alias{SingleCellAssay_Pvalues,Rscexv-method}
\title{description of function SingleCellAssay_Pvalues}
\usage{
SingleCellAssay_Pvalues(obj, ofile = "Significant_genes.csv")
}
\arguments{
\item{obj}{the Rscexv object}

\item{ofile}{the outfile default="Significant_genes.csv"}
}
\description{
USes the SingleCellAssay package to calculate p values.
}

