% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_this_color.R
\docType{methods}
\name{this.color}
\alias{this.color}
\alias{this.color,Rscexv-method}
\title{description of function createRFgrouping_col}
\usage{
this.color(x, name = NULL)
}
\arguments{
\item{x}{the Rscexv object}

\item{name}{the name of data column or empty if you want the last auto color}
}
\description{
Get the actual color range from the object
}

