% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvalues_write.stats.R
\docType{methods}
\name{write.stats}
\alias{write.stats}
\alias{write.stats,Rscexv-method}
\title{description of function write.stats}
\usage{
write.stats(stats = NULL, file = "lin_lang_stats.xls")
}
\arguments{
\item{stats}{the lin lang list default= NULL}

\item{file}{the outfile default='lin_lang_stats.xls'}
}
\description{
write a statistics table from the lin lang list
}

