% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_z.score.PCR.mad.R
\docType{methods}
\name{z.score.PCR.mad}
\alias{z.score.PCR.mad}
\alias{z.score.PCR.mad,Rscexv-method}
\title{description of function z.score.PCR.mad}
\usage{
z.score.PCR.mad(dataObj)
}
\arguments{
\item{dataObj}{the Rscexv object}
}
\description{
This normlization method is re-implemented from the MAST package.
}

