% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coexpression.R
\docType{methods}
\name{coexpressGenes}
\alias{coexpressGenes}
\alias{coexpressGenes,Rscexv-method}
\title{description of function coexpressGenes}
\usage{
coexpressGenes(dataObj)
}
\arguments{
\item{dataObj}{the Rscexv object}
}
\description{
calculates the coexpression for all genes in the data set
}

