% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coexpression.R
\docType{methods}
\name{coexpressGenes}
\alias{coexpressGenes}
\alias{coexpressGenes,Rscexv-method}
\title{description of function coexpressGenes}
\usage{
coexpressGenes(dataObj, grouping = NULL)
}
\arguments{
\item{dataObj}{the Rscexv object}

\item{grouping}{the column in the samples table that describes the grouping to use}
}
\description{
calculates the coexpression for all genes in all groups in the data set
}

