% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\docType{methods}
\name{complexHeatmap}
\alias{complexHeatmap}
\alias{complexHeatmap,Rscexv-method}
\title{description of function plot.beans}
\usage{
complexHeatmap(x, ofile, colGroups = NULL, rowGroups = NULL,
  colColors = NULL, rowColors = NULL, pdf = FALSE, subpath = "",
  main = "")
}
\arguments{
\item{x}{the Rscexv object}

\item{ofile}{the outfile to create in the x@outpath folder}

\item{colGroups}{columns in the samples table to use to order the data (first == order)}

\item{rowGroups}{rows in the annotation table to use to color the heatmap rows (first == order)}

\item{colColors}{a named list of column color vectors}

\item{rowColors}{a named list of row color vectors}

\item{pdf}{export as pdf (default = FALSE)}

\item{subpath}{the subpath for the plots (default = '')}
}
\description{
plot the PCR heatmap using the heatmap.3 function included in this package
}

