% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\docType{methods}
\name{plot.beans}
\alias{plot.beans}
\alias{plot.beans,Rscexv-method}
\title{description of function plot.beans}
\usage{
plot.beans(x, groups.n, clus, plot.neg = TRUE, mv = -20, subpath = "",
  names = NULL, col = NULL)
}
\arguments{
\item{x}{the Rscexv object}

\item{groups.n}{TEXT MISSING}

\item{clus}{TEXT MISSING}

\item{plot.neg}{TEXT MISSING default=TRUE}

\item{mv}{TEXT MISSING default=-20}

\item{subpath}{the subpath for the plots (default = '')}

\item{names}{a vector of grou names (default NULL)}

\item{col}{a colour vector for colouring the beans (default=NULL -> rainbow )}
}
\description{
plot the beanplots
}

